test keyboard

+++ screen.py
@@ -1053,50 +1053,59 @@ class KlipperScreen(Gtk.Window):
         self.show_panel('job_status', "job_status", "Print Status", 2)
         self.base_panel.show_heaters(True)
 
-    def show_keyboard(self, widget=None):
+    def show_keyboard(self, widget=None, event=None, entry=None):
         if self.keyboard is not None:
             return
 
-        env = os.environ.copy()
-        usrkbd = "/home/pi/.matchbox/keyboard.xml"
-        if os.path.isfile(usrkbd):
-            env["MB_KBD_CONFIG"] = usrkbd
-        else:
-            env["MB_KBD_CONFIG"] = "ks_includes/locales/keyboard.xml"
-        p = subprocess.Popen(["matchbox-keyboard", "--xid"], stdout=subprocess.PIPE,
-                             stderr=subprocess.PIPE, env=env)
+        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
+        box.set_size_request(self.gtk.get_content_width(), self.gtk.get_keyboard_height())
 
-        xid = int(p.stdout.readline())
-        logging.debug("XID %s" % xid)
-        logging.debug("PID %s" % p.pid)
-        keyboard = Gtk.Socket()
+        if self._config.get_main_config().getboolean("use-matchbox-keyboard", False):
+            env = os.environ.copy()
+            usrkbd = os.path.expanduser("~/.matchbox/keyboard.xml")
+            if os.path.isfile(usrkbd):
+                env["MB_KBD_CONFIG"] = usrkbd
+            else:
+                env["MB_KBD_CONFIG"] = "ks_includes/locales/keyboard.xml"
+            p = subprocess.Popen(["matchbox-keyboard", "--xid"], stdout=subprocess.PIPE,
+                                 stderr=subprocess.PIPE, env=env)
+            xid = int(p.stdout.readline())
+            logging.debug(f"XID {xid}")
+            logging.debug(f"PID {p.pid}")
 
-        action_bar_width = self.gtk.get_action_bar_width()
+            keyboard = Gtk.Socket()
+            box.get_style_context().add_class("keyboard_matchbox")
+            box.pack_start(keyboard, True, True, 0)
+            self.base_panel.get_content().pack_end(box, False, False, 0)
 
-        box = Gtk.VBox()
-        box.set_vexpand(False)
-        box.set_size_request(self.width - action_bar_width, self.keyboard_height)
-        box.add(keyboard)
+            self.show_all()
+            keyboard.add_id(xid)
 
-        self.base_panel.get_content().pack_end(box, False, 0, 0)
-
-        self.show_all()
-        keyboard.add_id(xid)
-        keyboard.show()
-
+            self.keyboard = {
+                "box": box,
+                "process": p,
+                "socket": keyboard
+            }
+            return
+        if entry is None:
+            logging.debug("Error: no entry provided for keyboard")
+            return
+        box.get_style_context().add_class("keyboard_box")
+        box.add(Keyboard(self, self.remove_keyboard, entry=entry))
         self.keyboard = {
-            "box": box,
-            # "panel": cur_panel.get(),
-            "process": p,
-            "socket": keyboard
+            "entry": entry,
+            "box": box
         }
+        self.base_panel.get_content().pack_end(box, False, False, 0)
+        self.base_panel.get_content().show_all()
 
-    def remove_keyboard(self, widget=None):
+    def remove_keyboard(self, widget=None, event=None):
         if self.keyboard is None:
             return
 
+        if 'process' in self.keyboard:
+            os.kill(self.keyboard['process'].pid, signal.SIGTERM)
         self.base_panel.get_content().remove(self.keyboard['box'])
-        os.kill(self.keyboard['process'].pid, signal.SIGTERM)
         self.keyboard = None
 
     def change_cursor(self, cursortype=None):
